==> notebooks/02_m2_clean_raw_data.ipynb <==
import pandas as pd
import numpy as np
from pathlib import Path

DATA_RAW = Path("../data/raw")
DATA_INTERIM = Path("../data/interim")
DATA_INTERIM.mkdir(parents=True, exist_ok=True)


product = pd.read_csv(DATA_RAW / "product.csv")
sales = pd.read_csv(DATA_RAW / "sales.csv")
discount = pd.read_csv(DATA_RAW / "discount.csv")
marketing = pd.read_csv(DATA_RAW / "marketing.csv")

print(product.shape, sales.shape, discount.shape, marketing.shape)


product.head()


sales.head()



discount.head()


marketing.head()


dup_count = sales.duplicated(
    subset=["week", "store_id", "product_id"]
).sum()

dup_count


sales_clean = sales.drop_duplicates(
    subset=["week", "store_id", "product_id"]
).copy()

sales_clean.shape


valid_products = set(product["product_id"])

invalid_rows = ~sales_clean["product_id"].isin(valid_products)
invalid_rows.sum()




sales_clean = sales_clean[~invalid_rows].copy()
sales_clean.shape


upper_cap = sales_clean["units_sold"].quantile(0.99)
sales_clean["units_sold"] = np.minimum(
    sales_clean["units_sold"], upper_cap
)

sales_clean["units_sold"].describe()



discount.duplicated(
    subset=["week", "store_id", "product_id"]
).sum()


marketing.duplicated(
    subset=["week", "store_id"]
).sum()


product.to_csv(DATA_INTERIM / "product_clean.csv", index=False)
sales_clean.to_csv(DATA_INTERIM / "sales_clean.csv", index=False)
discount.to_csv(DATA_INTERIM / "discount_clean.csv", index=False)
marketing.to_csv(DATA_INTERIM / "marketing_clean.csv", index=False)
